# docs/service/catalog-service/api.yaml
openapi: 3.0.3
info:
  title: catalog-service API
  version: "0.1"
  description: |
    Quản lý Catalog — Brand, Category, AttributeTemplate, Product (SPU), Variant (SKU).
    Upstream thuần túy: không consume event, chỉ publish.

    **Auth:**
    - `/api/**` — public (không cần token)
    - `/api/admin/**` — yêu cầu role `ADMIN` (bearer token từ oauth2-service)
servers:
  - url: http://localhost:8082
    description: Local dev
...
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
          example: null

    # ── Brand ─────────────────────────────────────────────────────────────────
...
    InputType:
      type: string
      enum: [SINGLE_SELECT, MULTI_SELECT, TEXT, NUMBER, BOOLEAN, DATE]
      description: >
        Bất biến. SINGLE/MULTI_SELECT có option (giá trị lưu là optionId). DATE lưu chuỗi ISO rút gọn
        (yyyy | yyyy-MM | yyyy-MM-dd). Quy tắc từng kiểu: analysis.md §5.3.1.
    AttributeOptionStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    AttributeTemplateStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    AttributeOptionResponse:
      type: object
      required: [id, value, displayValue, status, sortOrder]
      properties:
        id:
          type: string
          description: ULID
          example: "01HX4KZ7B3QNMR2VFPJ6W8Y9T"
        value:
          type: string
          description: Mã ổn định, bất biến, duy nhất trong template
          example: "black"
        displayValue:
          type: string
          example: "Đen"
        status:
          $ref: '#/components/schemas/AttributeOptionStatus'
        sortOrder:
          type: integer
          description: Thứ tự chuẩn trong template
    AttributeTemplateResponse:
      type: object
      required: [id, name, displayName, inputType, status, usable, options]
      description: >
        Chi tiết template (admin). Không tự áp dụng — phải được gán vào danh mục lá. `options` gồm MỌI option (kể cả
        đã tắt) theo `sortOrder`.
      properties:
        id: { type: string, description: ULID }
        name: { type: string, description: "Mã bất biến (snake_case)", example: "color" }
        displayName: { type: string, example: "Màu sắc" }
        hint: { type: string, nullable: true, description: "Gợi ý cách nhập cho seller (≤ 200)" }
        inputType:
          $ref: '#/components/schemas/InputType'
        unit: { type: string, nullable: true, description: "Chỉ NUMBER, bất biến", example: "GB" }
        status:
          $ref: '#/components/schemas/AttributeTemplateStatus'
        usable:
          type: boolean
          description: ACTIVE và (không phải SELECT hoặc còn ≥ 1 option ACTIVE) — còn được chọn mới không
        options:
          type: array
          items:
            $ref: '#/components/schemas/AttributeOptionResponse'
    AttributeTemplateSummaryPage:
      type: object
      required: [items, total]
      properties:
        total: { type: integer, format: int64 }
        items:
          type: array
          items:
            type: object
            required: [id, name, displayName, inputType, status, optionCount]
            properties:
              id: { type: string }
              name: { type: string }
              displayName: { type: string }
              inputType:
                $ref: '#/components/schemas/InputType'
              unit: { type: string, nullable: true }
              status:
                $ref: '#/components/schemas/AttributeTemplateStatus'
              optionCount: { type: integer }
    CreateAttributeTemplateRequest:
      type: object
      required: [name, displayName, inputType]
      properties:
        name:
          type: string
          description: Bất biến, duy nhất (kể cả template đã tắt)
          example: "ram"
        displayName:
          type: string
          example: "RAM"
        hint:
          type: string
          maxLength: 200
          nullable: true
        inputType:
          $ref: '#/components/schemas/InputType'
        unit:
          type: string
          maxLength: 20
          nullable: true
          description: Chỉ khi inputType=NUMBER (ngược lại 422 UNIT_ONLY_FOR_NUMBER); bất biến
          example: "GB"
    UpdateAttributeTemplateRequest:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          example: "Màu sắc (mới)"
        hint:
          type: string
          maxLength: 200
          nullable: true
    AddAttributeOptionRequest:
      type: object
      required: [value, displayValue]
      properties:
        value:
          type: string
          description: Mã ổn định (snake_case), bất biến, duy nhất trong template (không phân biệt hoa/thường)
          example: "black"
        displayValue:
          type: string
          description: Nhãn hiển thị
          example: "Đen"

  # ── Error responses ─────────────────────────────────────────────────────────
  responses:
    400:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: "VALIDATION_ERROR" }
                  message: { type: string }
    404:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
    409:
      description: Conflict (duplicate, in-use)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
    422:
      description: Business rule violation
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: "INPUT_TYPE_NOT_SELECT" }
                  message: { type: string }

    # ── Product ───────────────────────────────────────────────────────────────
...
paths:
...
  /api/admin/attribute-templates:
    get:
      tags: [AttributeTemplate (Admin)]
      summary: Danh sách template (quản trị)
      description: Mọi trạng thái; lọc + phân trang; trả bản tóm tắt (không kèm option, có số option).
      operationId: listAttributeTemplates
      parameters:
        - { name: keyword,   in: query, required: false, schema: { type: string }, description: "Tìm trong name/displayName, không phân biệt hoa/thường" }
        - { name: inputType, in: query, required: false, schema: { $ref: '#/components/schemas/InputType' } }
        - { name: status,    in: query, required: false, schema: { $ref: '#/components/schemas/AttributeTemplateStatus' } }
        - { name: page,      in: query, required: false, schema: { type: integer, default: 0, minimum: 0 } }
        - { name: size,      in: query, required: false, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/AttributeTemplateSummaryPage'
    post:
      tags: [AttributeTemplate (Admin)]
      summary: Tạo attribute template
      description: |
        `name`, `inputType`, `unit` bất biến sau khi tạo. `unit` chỉ hợp lệ với `inputType=NUMBER`.
        Trả về id của template mới.
      operationId: createAttributeTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttributeTemplateRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { type: string, description: "templateId (ULID)" }
        "400":
          $ref: '#/components/responses/400'
        "409":
          $ref: '#/components/responses/409'
          # ErrorCode: TEMPLATE_NAME_EXISTS
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: UNIT_ONLY_FOR_NUMBER | HINT_TOO_LONG
  /api/admin/attribute-templates/{id}:
    get:
      tags: [AttributeTemplate (Admin)]
      summary: Chi tiết template (quản trị)
      description: Đủ thuộc tính + MỌI option (kể cả đã tắt) theo `sortOrder`, cờ `usable`. Không kèm danh mục đang gán.
      operationId: getAttributeTemplate
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/AttributeTemplateResponse'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: TEMPLATE_NOT_FOUND
    put:
      tags: [AttributeTemplate (Admin)]
      summary: Sửa tên hiển thị và gợi ý
      description: Chỉ `displayName`, `hint` sửa được; `name`, `inputType`, `unit` bất biến.
      operationId: updateAttributeTemplate
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAttributeTemplateRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "400":
          $ref: '#/components/responses/400'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: TEMPLATE_NOT_FOUND
        "409":
          $ref: '#/components/responses/409'
          # optimistic lock — template vừa bị Admin khác sửa, tải lại rồi thử lại
    delete:
      tags: [AttributeTemplate (Admin)]
      summary: Tắt template (soft)
      description: |
        Chuyển sang INACTIVE, **không guard**. Không sửa assignment/sản phẩm/variant đang tham chiếu — chỉ chặn lựa chọn
        mới. Thuộc tính bắt buộc của danh mục tự hết hiệu lực khi template không dùng được.
      operationId: deactivateAttributeTemplate
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: TEMPLATE_NOT_FOUND
  /api/admin/attribute-templates/{id}/activate:
    post:
      tags: [AttributeTemplate (Admin)]
      summary: Bật lại template
      description: Không guard. Ưu tiên bật lại thay vì tạo template mới (`name` duy nhất trên mọi row).
      operationId: activateAttributeTemplate
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: TEMPLATE_NOT_FOUND
  /api/admin/attribute-templates/{id}/options:
    post:
      tags: [AttributeTemplate (Admin)]
      summary: Thêm option
      description: |
        Chỉ template SINGLE_SELECT/MULTI_SELECT đang ACTIVE. `value` duy nhất trong template (không phân biệt hoa/thường).
        Option mới xếp cuối. Trả về id của option mới.
      operationId: addAttributeOption
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: AttributeTemplate ULID }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAttributeOptionRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { type: string, description: "optionId (ULID)" }
        "400":
          $ref: '#/components/responses/400'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: TEMPLATE_NOT_FOUND
        "409":
          $ref: '#/components/responses/409'
          # ErrorCode: OPTION_VALUE_EXISTS
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: INPUT_TYPE_NOT_SELECT | TEMPLATE_INACTIVE
  /api/admin/attribute-templates/{id}/options/order:
    put:
      tags: [AttributeTemplate (Admin)]
      summary: Sắp lại thứ tự chuẩn của option
      description: Gửi MỌI optionId của template theo thứ tự mong muốn, mỗi option đúng 1 lần.
      operationId: reorderAttributeOptions
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [optionIds]
              properties:
                optionIds:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: TEMPLATE_NOT_FOUND | OPTION_NOT_FOUND
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: INVALID_OPTION_ORDER
  /api/admin/attribute-templates/{id}/options/{optionId}:
    put:
      tags: [AttributeTemplate (Admin)]
      summary: Sửa nhãn option
      description: |
        Chỉ `displayValue` sửa được (`value` bất biến). Không guard theo trạng thái — sửa được kể cả option
        đã tắt hoặc template đã tắt (sửa mô tả không phải "thêm mới"/"chọn dùng", không bị trạng thái chi
        phối). Áp ngay mọi nơi đang dùng. Dùng để sửa cách gọi — đổi NGHĨA thì tạo option mới + tắt option cũ.
      operationId: updateAttributeOption
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: AttributeTemplate ULID }
        - { name: optionId, in: path, required: true, schema: { type: string }, description: AttributeOption ULID }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayValue]
              properties:
                displayValue:
                  type: string
                  example: "Đen Matte"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "400":
          $ref: '#/components/responses/400'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: TEMPLATE_NOT_FOUND | OPTION_NOT_FOUND
    delete:
      tags: [AttributeTemplate (Admin)]
      summary: Tắt option (soft)
      description: |
        Chuyển sang INACTIVE, **không guard** (kể cả khi đang có variant dùng, kể cả khi template đã tắt). Sản phẩm/variant
        đang dùng giữ nguyên; chỉ chặn lựa chọn mới.
      operationId: deactivateAttributeOption
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: AttributeTemplate ULID }
        - { name: optionId, in: path, required: true, schema: { type: string }, description: AttributeOption ULID }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: TEMPLATE_NOT_FOUND | OPTION_NOT_FOUND
  /api/admin/attribute-templates/{id}/options/{optionId}/activate:
    post:
      tags: [AttributeTemplate (Admin)]
      summary: Bật lại option
      description: Không guard. Ưu tiên bật lại option cũ thay vì tạo option mới cùng `value` (tránh 2 ID cho 1 khái niệm).
      operationId: activateAttributeOption
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: AttributeTemplate ULID }
        - { name: optionId, in: path, required: true, schema: { type: string }, description: AttributeOption ULID }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: TEMPLATE_NOT_FOUND | OPTION_NOT_FOUND

  # ── Product (Public) ────────────────────────────────────────────────────────
...